pyproject.toml
[tool.poetry]
name = "InstituteQnA"
version = "0.1.0"
description = "Poetry project for InstituteQnA"
authors = ["Aniket Jawale <aniketjawale17@gmail.com>"]
packages = [{ include = "institute_qna" }]

[tool.poetry.dependencies]
python = "^3.12"
aiohappyeyeballs = "2.6.1"
aiohttp = "3.13.2"
aiosignal = "1.4.0"
annotated-types = "0.7.0"
anyio = "4.11.0"
attrs = "25.4.0"
azure-core = "1.36.0"
azure-search-documents = "^11.5.2"
azure-storage-blob = "^12.27.1"
beautifulsoup4 = "4.14.2"
bs4 = "0.0.2"
build = "1.3.0"
cachecontrol = "0.14.3"
cachetools = "6.2.1"
certifi = "2025.10.5"
cffi = "2.0.0"
charset-normalizer = "3.4.4"
cleo = "2.1.0"
crashtest = "0.4.1"
dataclasses-json = "0.6.7"
distlib = "0.4.0"
distro = "1.9.0"
dulwich = "0.24.8"
fastjsonschema = "2.21.2"
filelock = "3.20.0"
findpython = "0.7.0"
frozenlist = "1.8.0"
google-ai-generativelanguage = "0.6.15"
google-api-core = "2.25.2"
google-api-python-client = "2.187.0"
google-auth = "2.43.0"
google-auth-httplib2 = "0.2.1"
google-generativeai = "0.8.5"
googleapis-common-protos = "1.71.0"
grpcio = "1.76.0"
grpcio-status = "1.71.2"
h11 = "0.16.0"
httpcore = "1.0.9"
httplib2 = "0.31.0"
httpx = "0.28.1"
httpx-sse = "0.4.3"
idna = "3.11"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jaraco-context = "6.0.1"
jaraco-functools = "4.3.0"
jiter = "0.11.1"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
keyring = "25.6.0"
langchain-classic = "1.0.0"
langchain-community = "0.4.1"
langchain-core = "1.0.3"
langchain-text-splitters = "1.0.0"
langsmith = "0.4.41"
marshmallow = "3.26.1"
more-itertools = "10.8.0"
msgpack = "1.1.2"
multidict = "6.7.0"
mypy-extensions = "1.1.0"
numpy = "2.3.4"
openai = "2.7.1"
orjson = "3.11.4"
packaging = "25.0"
pbs-installer = "2025.10.31"
pkginfo = "1.12.1.2"
platformdirs = "4.5.0"
poetry = "2.2.1"
poetry-core = "2.2.1"
propcache = "0.4.1"
proto-plus = "1.26.1"
protobuf = "5.29.5"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.2"
pycparser = "2.23"
pydantic = "2.12.4"
pydantic-settings = "2.11.0"
pydantic-core = "2.41.5"
pyparsing = "3.2.5"
pyproject-hooks = "1.2.0"
python-dotenv = "1.2.1"
pyyaml = "6.0.3"
rapidfuzz = "3.14.3"
requests = "2.32.5"
requests-toolbelt = "1.0.0"
rsa = "4.9.1"
shellingham = "1.5.4"
sniffio = "1.3.1"
soupsieve = "2.8"
sqlalchemy = "2.0.44"
tenacity = "9.1.2"
tomlkit = "0.13.3"
tqdm = "4.67.1"
trove-classifiers = "2025.9.11.17"
typing-inspect = "0.9.0"
typing-inspection = "0.4.2"
typing-extensions = "4.15.0"
uritemplate = "4.2.0"
urllib3 = "2.5.0"
virtualenv = "20.35.4"
xattr = "1.3.0"
yarl = "1.22.0"
zstandard = "0.25.0"
fastapi = "^0.121.0"
uvicorn = "^0.38.0"
markdownify = "^1.2.0"
langchain = "^1.0.4"
pypdf = "^6.1.3"
langchain-openai = "^1.0.2"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pillow = "^12.0.0"
pdfminer = "^20191125"
pdfminer-six = "^20251107"
pi-heif = "^1.1.1"
langchain-chroma = "^1.0.0"
pandas = "^2.3.3"
azure-ai-documentintelligence = "^1.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^24.1.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
